clipsazo_api/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0